requirements.txt
discord.py
rgbprint
httpx
pillow